Enforce index uniqueness inside the Generator struct
This is the first step to allow PRSS to generate only one part of randomness, shared with one helper. The motivating issue is #1029. To support that, we need to enforce uniqueness checks at the lower layer because each generator now can be called at different times.

This change moves the `used` variable to `Generator` struct. This required some rethinking of `generate` method parameter types - instead of raw `u128`, it now takes `PrssIndexu128`. Mainly for debugging purposes, to be able to print "index:offset" when a duplicate index is used.