Change Validator trait to use UpgradedContext
This is a minor change that does not affect a lot of code.
The reason for making it is to make trait bounds easier
for the upcoming PRF eval upgrade where the validator
bounds need to be expressed in a form of upgraded context.

It is also easier to reason about because at the time
when validator is created, the `UpgradableContext` is gone