Use Upgradable trait in MAC module
When working on vectorizing malicious shares,
we noticed that the existing upgrade API does not work
for vectorized shares. It either required contexts to be
vectorization-aware which we tried to avoid up to this point,
or a different mechanism to specify how upgrades occur
for vectorized shares.

For example: `Replicated<F, N>` when upgraded, gets
transformed to `MaliciousReplicated<F, N>` and different
`N` values, require different bounds on functions.

That was a blocker for PRF evaluation upgrade and the solution
we came up with was to introduce `Upgradable` trait.

The functionality of this trait is very similar to `UpgradeToMalicious`
and `Upgrade` but upgrade now is bound explicitly to `RecordId`. This
optimizes for protocols readability. Malicious upgrades in tests had
to find a way around this limitation, and it is a bit hacky. I don't
think it is significantly different from what we've been doing before
though.